requirements.txt
Django>=1.6
django_compressor>=1.4